Updating Calendar DatePicker for Task creation
